fix(saving): CodeRabbit 리뷰 반영 - null 검증/동시성/오버플로 (#220)
- selections 리스트에 null 항목이 있으면 NPE로 500이 나던 문제를 List<@NotNull ...>로 차단
- 동시 PUT 요청이 upsert/delete 잠금 순서가 엇갈리며 교착되는 걸 막기 위해,
  저장 시작 시 monthly_spending_analyses 부모 행에 SELECT ... FOR UPDATE로 잠금 추가
  (실제 DB 동시성 통합 테스트는 이 환경에 DB 접근이 없어 대신 Mockito InOrder로
  락 -> upsert -> delete 호출 순서만 검증)
- totalMonthlyReductionTarget/totalWeeklyExpectedSaving을 Integer -> Long으로 변경해
  다수 카테고리 합산 시 int 오버플로 가능성 제거
- 잘못된 yearMonth 테스트에 verifyNoInteractions 추가

+++ backend/src/main/java/com/tripass/saving/dto/MissionSelectionsResponseDto.java
@@ -8,8 +8,8 @@ import java.util.List;
  */
 public record MissionSelectionsResponseDto(
         Integer selectedMissionCount,
-        Integer totalMonthlyReductionTarget,
-        Integer totalWeeklyExpectedSaving,
+        Long totalMonthlyReductionTarget,
+        Long totalWeeklyExpectedSaving,
         List<MissionSelectionResponseDto> selections
 ) {
 }
